calyx_ir
Enum Control

pub enum Control {
    Seq(Seq),
    Par(Par),
    If(If),
    While(While),
    Repeat(Repeat),
    Invoke(Invoke),
    Enable(Enable),
    Empty(Empty),
    Static(StaticControl),
}

Control AST nodes.

Variants

Seq(Seq)

Represents sequential composition of control statements.

Par(Par)

Represents parallel composition of control statements.

If(If)

Standard imperative if statement

While(While)

Standard imperative while statement

Repeat(Repeat)

Standard repeat control statement

Invoke(Invoke)

Invoke a sub-component with the given port assignments

Enable(Enable)

Runs the control for a list of subcomponents.

Empty(Empty)

Control statement that does nothing.

Static(StaticControl)

Static Control

Implementations

impl Control

pub fn empty() -> Self

Convience constructor for empty.

pub fn seq(stmts: Vec<Control>) -> Self

Convience constructor for seq.

pub fn par(stmts: Vec<Control>) -> Self

Convience constructor for par.

pub fn enable(group: RRC<Group>) -> Self

Convience constructor for enable.

pub fn static_enable(group: RRC<StaticGroup>) -> Self

Convience constructor for enable.

pub fn invoke( comp: RRC<Cell>, inputs: Vec<(Id, RRC<Port>)>, outputs: Vec<(Id, RRC<Port>)>, ) -> Self

Convience constructor for invoke.

pub fn if_( port: RRC<Port>, cond: Option<RRC<CombGroup>>, tbranch: Box<Control>, fbranch: Box<Control>, ) -> Self

Convience constructor for if

pub fn while_( port: RRC<Port>, cond: Option<RRC<CombGroup>>, body: Box<Control>, ) -> Self

Convience constructor for while

pub fn repeat(num_repeats: u64, body: Box<Control>) -> Self

Convience constructor for dynamic repeat

pub fn get_attribute<A>(&self, attr: A) -> Option<u64>
where A: Into<Attribute>,

Returns the value of an attribute if present

pub fn has_attribute<A>(&self, attr: A) -> bool
where A: Into<Attribute>,

Returns true if the node has a specific attribute

pub fn is_static(&self) -> bool

pub fn is_empty(&self) -> bool

pub fn get_latency(&self) -> Option<u64>

pub fn take_control(&mut self) -> Control

Replaces &mut self with an empty control statement, and returns self

pub fn take_static_control(&mut self) -> StaticControl

Replaces &mut self with an empty control statement, and returns StaticControl of self. Note that this only works on Control that is static