Enum calyx_ir::Guard

pub enum Guard<T> {
    Or(Box<Guard<T>>, Box<Guard<T>>),
    And(Box<Guard<T>>, Box<Guard<T>>),
    Not(Box<Guard<T>>),
    True,
    CompOp(PortComp, RRC<Port>, RRC<Port>),
    Port(RRC<Port>),
    Info(T),
}

An assignment guard which has pointers to the various ports from which it reads.

Variants

Or(Box<Guard<T>>, Box<Guard<T>>)

Represents c1 || c2.

And(Box<Guard<T>>, Box<Guard<T>>)

Represents c1 && c2.

Not(Box<Guard<T>>)

Represents !c1

True

The constant true

CompOp(PortComp, RRC<Port>, RRC<Port>)

Comparison operator.

Port(RRC<Port>)

Uses the value on a port as the condition. Same as p1 == true

Info(T)

Other types of information.

Implementations

impl<T> Guard<T>

pub fn is_true(&self) -> bool

Returns true definitely Guard::True. Returning false does not mean that the guard is not true.

pub fn is_false(&self) -> bool

Checks if the guard is always false. Returning false does not mean that the guard is not false.

pub fn is_not_done(&self, cell_name: &Id) -> bool

returns true if the self is !cell_name, false otherwise.

pub fn update<F>(&mut self, upd: F)where
F: FnOnce(Guard<T>) -> Guard<T>,

Update the guard in place. Replaces this guard with upd(self). Uses std::mem::take for the in-place update.

pub fn op_str(&self) -> String

Return the string corresponding to the guard operation.

pub fn port(p: RRC<Port>) -> Self

pub fn and(self, rhs: Guard<T>) -> Selfwhere
T: Eq,

pub fn or(self, rhs: Guard<T>) -> Selfwhere
T: Eq,

pub fn eq(self, other: Guard<T>) -> Selfwhere
T: Debug,

pub fn neq(self, other: Guard<T>) -> Selfwhere
T: Debug,

pub fn le(self, other: Guard<T>) -> Selfwhere
T: Debug,

pub fn lt(self, other: Guard<T>) -> Selfwhere
T: Debug,

pub fn ge(self, other: Guard<T>) -> Selfwhere
T: Debug,

pub fn gt(self, other: Guard<T>) -> Selfwhere
T: Debug,

pub fn all_ports(&self) -> Vec<RRC<Port>>

Returns all the ports used by this guard.

impl<T> Guard<T>

Helper functions for the guard.

pub fn for_each<F>(&mut self, f: &mut F)where
F: FnMut(RRC<Port>) -> Option<Guard<T>>,

Mutates a guard by calling f on every leaf in the guard tree and replacing the leaf with the guard that f returns.

pub fn for_each_info<F>(&mut self, f: &mut F)where
F: FnMut(&mut T) -> Option<Guard<T>>,

runs f(info) on each Guard::Info in guard. if f(info) = Some(result)replaces interval with result. iff(info) = None does nothing.

pub fn check_for_each_info<F>(&self, f: &mut F) -> Result<(), Error>where
F: Fn(&T) -> Result<(), Error>,

runs f(info) on each info in guard. f should return Result<(), Error>, meaning that it essentially does nothing if the f returns OK(()), but returns an appropraite error otherwise

impl Guard<StaticTiming>

pub fn add_interval(&mut self, timing_interval: StaticTiming)

updates self -> self & interval