Struct calyx_ir::rewriter::Rewriter

pub struct Rewriter<'a> { /* private fields */ }

A structure to track rewrite maps for ports. Stores both cell rewrites and direct port rewrites. Attempts to apply port rewrites first before trying the cell rewrite.

Implementations

impl<'a> Rewriter<'a>

pub fn new(cell_map: &'a RewriteMap<Cell>, port_map: &'a PortRewriteMap) -> Self

pub fn get_cell_rewrite(&self, cell: &Id) -> Option<RRC<Cell>>

Return the rewrite for a cell

pub fn get(&self, port_ref: &RRC<Port>) -> Option<RRC<Port>>

Get any port rewrite defined for the given port.

pub fn rewrite_invoke(
&self,
inv: &mut Invoke,
comb_group_map: &RewriteMap<CombGroup>
)

Rewrite a invoke node using a RewriteMapir::Cell and a RewriteMapir::CombGroup

pub fn rewrite_static_invoke(&self, inv: &mut StaticInvoke)

Rewrite a static invoke node using a RewriteMapir::Cell and a RewriteMapir::CombGroup

pub fn rewrite_static_control(
&self,
sc: &mut StaticControl,
static_group_map: &RewriteMap<StaticGroup>
)

Given a control program, rewrite all uses of cells, groups, and comb groups using the given rewrite maps.

pub fn rewrite_control(
&self,
c: &mut Control,
group_map: &RewriteMap<Group>,
comb_group_map: &RewriteMap<CombGroup>,
static_group_map: &RewriteMap<StaticGroup>
)

Given a control program, rewrite all uses of cells, groups, and comb groups using the given rewrite maps.