calyx_ir::source_info

Struct SourceInfoTable

pub struct SourceInfoTable { /* private fields */ }

Implementations

impl SourceInfoTable

pub fn lookup_file_path(&self, file: FileId) -> &PathBuf

Looks up the path of the file with the given id.

Panics

Panics if the file id does not exist in the file map

pub fn lookup_position(&self, pos: PositionId) -> &SourceLocation

Looks up the source location of the position with the given id.

Panics

Panics if the position id does not exist in the position map

pub fn get_position(&self, pos: PositionId) -> Option<&SourceLocation>

Looks up the source location of the position with the given id. If no such position exists, returns None

pub fn iter_file_map(&self) -> impl Iterator<Item = (&FileId, &PathBuf)>

Iterate over the stored file map, returning a tuple of references to the file id and the path

pub fn iter_file_paths(&self) -> impl Iterator<Item = &PathBuf>

Iterate over the paths of all files in the file map

pub fn iter_file_ids(&self) -> impl Iterator<Item = FileId>

Iterate over all file ids in the file map

pub fn iter_position_map( &self, ) -> impl Iterator<Item = (&PositionId, &SourceLocation)>

Iterate over the stored position map, returning a tuple of references to the position id and the source location

pub fn iter_positions(&self) -> impl Iterator<Item = PositionId>

Iterate over all position ids in the position map

pub fn iter_source_locations(&self) -> impl Iterator<Item = &SourceLocation>

Iterate over the source locations in the position map

pub fn add_file(&mut self, file: FileId, path: PathBuf)

Adds a file to the file map with the given id

pub fn push_file(&mut self, path: PathBuf) -> FileId

Adds a file to the file map and generates a new file id for it. If you want to add a file with a specific id, use SourceInfoTable::add_file

pub fn add_position( &mut self, pos: PositionId, file: FileId, line: LineNum, endline: Option<LineNum>, )

pub fn push_position( &mut self, file: FileId, line: LineNum, endline: Option<LineNum>, ) -> PositionId

Adds a position to the position map and generates a new position id for it. If you want to add a position with a specific id, use SourceInfoTable::add_position

pub fn add_location(&mut self, id: MemoryLocationId, info: MemoryLocation)

pub fn get_variable_mapping( &self, pos: PositionId, ) -> Option<&HashMap<String, MemoryLocationId>>

Attempts to look up the variable mapping associated with a given position, if such a mapping exists

pub fn get_memory_location(&self, loc: &MemoryLocationId) -> &MemoryLocation

pub fn new_empty() -> SourceInfoTable

Creates a new empty source info table

pub fn new_minimal( files: impl IntoIterator<Item = (FileId, PathBuf)>, positions: impl IntoIterator<Item = (PositionId, FileId, LineNum, Option<LineNum>)>, ) -> Result<SourceInfoTable, SourceInfoTableError>

A wrapper function to construct a source a source map containing only files and positions. If an empty map is needed use SourceInfoTable::new_empty

pub fn new( files: impl IntoIterator<Item = (FileId, PathBuf)>, positions: impl IntoIterator<Item = (PositionId, FileId, LineNum, Option<LineNum>)>, locations: impl IntoIterator<Item = (MemoryLocationId, MemoryLocation)>, variable_assigns: impl IntoIterator<Item = (VariableAssignmentId, impl IntoIterator<Item = (String, MemoryLocationId)>)>, states: impl IntoIterator<Item = (PositionId, VariableAssignmentId)>, ) -> Result<SourceInfoTable, SourceInfoTableError>

pub fn serialize<W>(&self, f: W) -> Result<(), Error>
where W: Write,

pub fn get_position_string( &self, pos: PositionId, ) -> Result<String, SourceInfoTableError>

Attempt to lookup the line that a given position points to. Returns an error in cases when the position does not exist, the file is unavailable, or the file does not contain the indicated line.