calyx_ir

Struct FSM

pub struct FSM {
    pub attributes: Attributes,
    pub assignments: Vec<Vec<Assignment<Nothing>>>,
    pub transitions: Vec<Transition>,
    pub wires: SmallVec<[RRC<Port>; 2]>,
    /* private fields */
}

Fields
attributes: Attributes

Attributes for this FSM

assignments: Vec<Vec<Assignment<Nothing>>>

State indexes into assignments that are supposed to be enabled at that state

transitions: Vec<Transition>

State indexes into (potentially guarded) next states

wires: SmallVec<[RRC<Port>; 2]>
Implementations

impl FSM

pub fn name(&self) -> Id

Grants immutable access to the name of this cell.

pub fn new(name: Id) -> Self

Constructs a new FSM construct using a list of cases and a name

pub fn num_states(&self) -> u64

Gets the number of states in the FSM.

pub fn find<S>(&self, name: S) -> Option<RRC<Port>>
where S: Display, Id: PartialEq<S>,

Get a reference to the named hole if it exists.

pub fn get<S>(&self, name: S) -> RRC<Port>
where S: Display + Clone, Id: PartialEq<S>,

pub fn extend_fsm<A, T>(&mut self, assigns: A, transitions: T)
where A: IntoIterator<Item = Vec<Assignment<Nothing>>>, T: IntoIterator<Item = Transition>,

Extend the FSM with new transitions and assignments. Will panic if the lengths are not consistent.

pub fn extend_state_assignments<I>(&mut self, state: u64, assigns: I)
where I: IntoIterator<Item = Assignment<Nothing>>,

Extend the assignments that are supposed to be active at a given state.

pub fn get_called_port_parents<F>(&self, push_parent_name: F) -> Vec<Id>
where F: Fn(&mut Vec<Id>, &RRC<Port>),

Returns a list of names of the groups, cells, or other port parents used by the FSM. Requires as an argument a function that can in-place update a Vec with the name of kinds of port parents you care about.

pub fn merge_assignments(&self) -> Vec<Vec<(usize, Assignment<Nothing>)>>

Each element of the resulting Vec is a collection of assignments writing to the same destination port, where the usize value represents the FSM state at which the assignment should take place.