calyx_ir

Struct IdList

pub struct IdList<T: GetName>(/* private fields */);

A wrapper struct exposing an ordered collection of named entities within an RRC with deterministic iteration and constant-time look-up on names directly. The struct assumes that the name of an entity cannot change. Doing so will introduce incorrect results for look-ups.

Implementations

impl<T: GetName> IdList<T>

pub fn clear(&mut self)

Removes all elements from the collection

pub fn is_empty(&self) -> bool

Returns true if there are no elements in the list.

pub fn len(&self) -> usize

pub fn retain<F>(&mut self, f: F) -> u64
where F: FnMut(&RRC<T>) -> bool,

Keep only the elements in the collection which satisfy the given predicate and return the number of elements removed.

pub fn add(&mut self, item: RRC<T>)

Add a new element to the colleciton

pub fn remove<S>(&mut self, name: S) -> Option<RRC<T>>
where S: Into<Id>,

pub fn append(&mut self, items: impl Iterator<Item = RRC<T>>)

Add all elements to the collection

pub fn iter(&self) -> impl Clone + Iterator<Item = &RRC<T>>

Returns an iterator over immutable references

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut RRC<T>>

Returns an iterator over mutable references. Likely a pointless method as this is a collection of RRCs.

pub fn drain(&mut self) -> impl Iterator<Item = RRC<T>> + use<T>

Removes all elements from the collection and returns an iterator over the owned elements.

pub fn find<S>(&self, name: S) -> Option<RRC<T>>
where S: Into<Id>,

Returns the element indicated by the name, if present, otherwise None.