calyx_ir

Struct PortDef

pub struct PortDef<W> {
    pub width: W,
    pub direction: Direction,
    pub attributes: Attributes,
    /* private fields */
}

Definition of a port parameterized by a width type. Ports on Primitives can be parameteris and use Width. Ports on Components cannot be parameterized and therefore use u64.

Fields
width: W

The width of the port. .

direction: Direction

The direction of the port. Only allowed to be Direction::Input or Direction::Output.

attributes: Attributes

Attributes attached to this port definition

Implementations

impl<W> PortDef<W>

pub fn new( name: impl Into<Id>, width: W, direction: Direction, attributes: Attributes, ) -> PortDef<W>

pub fn name(&self) -> Id

Return the name of the port definition

impl PortDef<Width>

pub fn resolve( self, binding: &LinkedHashMap<Id, u64>, ) -> Result<PortDef<u64>, Error>

Given a map from names of parameters to their values, attempt to resolve this definition. Errors if there is no binding for a required parameter binding.