Struct calyx_ir::Printer

pub struct Printer;

Printer for the IR.

Implementations

impl Printer

pub fn format_at_attributes(attrs: &Attributes) -> String

Format attributes of the form @static(1). Returns the empty string if the attrs is empty.

pub fn format_attributes(attrs: &Attributes) -> String

Format attributes of the form <"static"=1>. Returns the empty string if the attrs is empty.

pub fn format_ports(ports: &[RRC<Port>]) -> String

Formats port definitions in signatures

pub fn format_port_def<W: Display>(port_defs: &[&PortDef<W>]) -> String

pub fn write_context<F: Write>( ctx: &Context, skip_primitives: bool, f: &mut F, ) -> Result<()>

Prints out the program context. If skip_primitives is true, the printer will skip printing primitives defined outside the source file.

pub fn write_externs<'a, F, I>( (path, prims): (&Path, I), f: &mut F, ) -> Result<()>
where F: Write, I: Iterator<Item = &'a Primitive>,

Formats and writes extern statements.

pub fn write_primitive<F: Write>( prim: &Primitive, indent: usize, f: &mut F, ) -> Result<()>

pub fn write_component<F: Write>(comp: &Component, f: &mut F) -> Result<()>

Formats and writes the Component to the formatter.

pub fn write_float_const<F: Write>( cell: &Cell, indent_level: usize, f: &mut F, ) -> Result<()>

pub fn write_cell<F: Write>( cell: &Cell, indent_level: usize, f: &mut F, ) -> Result<()>

Format and write a cell.

pub fn write_assignment<F: Write, T: Clone + ToString + Eq>( assign: &Assignment<T>, indent_level: usize, f: &mut F, ) -> Result<()>

Format and write an assignment.

pub fn assignment_to_str<T>(assign: &Assignment<T>) -> String
where T: ToString + Clone + Eq,

Convinience method to get string representation of ir::Assignment.

pub fn control_to_str(assign: &Control) -> String

Convinience method to get string representation of ir::Control.

pub fn write_comb_group<F: Write>( group: &CombGroup, indent_level: usize, f: &mut F, ) -> Result<()>

Format and write a combinational group.

pub fn write_group<F: Write>( group: &Group, indent_level: usize, f: &mut F, ) -> Result<()>

Format and write a group.

pub fn write_static_group<F: Write>( group: &StaticGroup, indent_level: usize, f: &mut F, ) -> Result<()>

Format and write a static group.

pub fn write_static_control<F: Write>( scontrol: &StaticControl, indent_level: usize, f: &mut F, ) -> Result<()>

Format and write a static control program

pub fn write_control<F: Write>( control: &Control, indent_level: usize, f: &mut F, ) -> Result<()>

Format and write a control program

pub fn guard_str<T>(guard: &Guard<T>) -> String
where T: Eq + ToString,

Generate a String-based representation for a guard.

pub fn port_to_str(port: &Port) -> String

Get the port access expression.

pub fn format_metadata(metadata: &Option<String>) -> String

Formats the top-level metadata if present

pub fn write_source_info_table<W: Write>( f: &mut W, metadata: &Option<SourceInfoTable>, ) -> Result<(), Error>