Struct calyx_ir::rewriter::Rewriter

pub struct Rewriter {
    pub port_map: PortRewriteMap,
    pub cell_map: RewriteMap<Cell>,
    pub group_map: RewriteMap<Group>,
    pub comb_group_map: RewriteMap<CombGroup>,
    pub static_group_map: RewriteMap<StaticGroup>,
}

A structure to track rewrite maps for ports. Stores both cell rewrites and direct port rewrites. Attempts to apply port rewrites first before trying the cell rewrite.

Fields
port_map: PortRewriteMap

Mapping from canonical names of ports to port instances

cell_map: RewriteMap<Cell>

Mapping from names of cells to cell instance.

group_map: RewriteMap<Group>

Mapping from names of groups to group instance.

comb_group_map: RewriteMap<CombGroup>

Mapping from names of combinational groups to combinational group instance.

static_group_map: RewriteMap<StaticGroup>

Mapping from names of static groups to static group instance.

Implementations

impl Rewriter

pub fn get_cell_rewrite(&self, cell: &Id) -> Option<RRC<Cell>>

Return the rewrite for a cell

pub fn get(&self, port_ref: &RRC<Port>) -> Option<RRC<Port>>

Get any port rewrite defined for the given port.

pub fn rewrite_guard<T>(&self, guard: &mut Guard<T>)

Rewrite assignments in a guard

pub fn rewrite_assign<T>(&self, assign: &mut Assignment<T>)

Rewrite an assignment

pub fn rewrite_invoke(&self, inv: &mut Invoke)

Rewrite a invoke node using a RewriteMapir::Cell and a RewriteMapir::CombGroup

pub fn rewrite_static_invoke(&self, inv: &mut StaticInvoke)

Rewrite a static invoke node using a RewriteMapir::Cell and a RewriteMapir::CombGroup

pub fn rewrite_static_control(&self, sc: &mut StaticControl)

Given a control program, rewrite all uses of cells, groups, and comb groups using the given rewrite maps.

pub fn rewrite_control(&self, c: &mut Control)

Given a control program, rewrite all uses of cells, groups, and comb groups using the given rewrite maps.

pub fn rewrite(&self, comp: &mut Component)

Rewrite the component using the given maps